Move meaning analysis into a separate crate
Let's have it cleanly separated from what would turn into macroexpander.
Semantic meaning analysis is pretty straightforward, but I really like
splitting the codebase into crates aggressively. Rust compilation times
are still abysmal so it makes sense.
