+++ src/libmeaning/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "libmeaning"
+version = "0.1.0"
+edition = "2018"
+authors = ["Sabre developers"]
+license = "MIT/Apache-2.0"
+
+[dependencies]
+libeval = { path = "../libeval", version = "0.1.0" }
+liblocus = { path = "../liblocus", version = "0.1.0" }
+libreader = { path = "../libreader", version = "0.1.0" }
