+++ src/libmeaning/README.md
@@ -1,0 +1,4 @@
+libmeaning
+==========
+
+Core Scheme semantics.
