+++ src/libmeaning/tests/meaning.rs
@@ -9,7 +9,7 @@
 //!
 //! This verifies that the basic semantics of Scheme is handled as expected.
 
-use libeval::meaning::{meaning, MeaningResult, Value};
+use libmeaning::{meaning, MeaningResult, Value};
 
 // - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 // Tested expanders and environments
@@ -24,28 +24,44 @@ use liblocus::diagnostics::{DiagnosticKind, Span};
 use libreader::intern_pool::InternPool;
 
 macro_rules! syntax {
-    ($pool:expr, $name:expr, $type:ty) => ({
+    ($pool:expr, $name:expr, $type:ty) => {{
         let name = $pool.intern($name);
-        (Variable { name, span: Span::new(0, 0) }, Box::new(<$type>::new(name)))
-    })
+        (
+            Variable {
+                name,
+                span: Span::new(0, 0),
+            },
+            Box::new(<$type>::new(name)),
+        )
+    }};
 }
 
 fn basic_scheme_environment(pool: &InternPool) -> Rc<Environment> {
     let keywords: Vec<(Variable, Box<Expander>)> = vec![
-        syntax!(pool, "quote",  QuoteExpander),
-        syntax!(pool, "begin",  BeginExpander),
-        syntax!(pool, "if",     IfExpander),
-        syntax!(pool, "set!",   SetExpander),
+        syntax!(pool, "quote", QuoteExpander),
+        syntax!(pool, "begin", BeginExpander),
+        syntax!(pool, "if", IfExpander),
+        syntax!(pool, "set!", SetExpander),
         syntax!(pool, "lambda", LambdaExpander),
     ];
     let imported_vars = [
-        Variable { name: pool.intern("car"), span: Span::new(0, 0) },
-        Variable { name: pool.intern("cdr"), span: Span::new(0, 0) },
-        Variable { name: pool.intern("cons"), span: Span::new(0, 0) },
+        Variable {
+            name: pool.intern("car"),
+            span: Span::new(0, 0),
+        },
+        Variable {
+            name: pool.intern("cdr"),
+            span: Span::new(0, 0),
+        },
+        Variable {
+            name: pool.intern("cons"),
+            span: Span::new(0, 0),
+        },
     ];
-    let global_vars = [
-        Variable { name: pool.intern("*global*"), span: Span::new(0, 0) },
-    ];
+    let global_vars = [Variable {
+        name: pool.intern("*global*"),
+        span: Span::new(0, 0),
+    }];
 
     let imported_env = Environment::new_imported(&imported_vars, keywords);
     let global_env = Environment::new_global(&global_vars, &imported_env);
@@ -447,7 +463,8 @@ impl TestCase {
     }
 
     fn constants<F>(mut self, generator: F) -> Self
-        where F: Fn(&InternPool) -> Vec<Value> + 'static
+    where
+        F: Fn(&InternPool) -> Vec<Value> + 'static,
     {
         assert!(self.constant_generator.is_none(), "don't set constants twice");
         self.constant_generator = Some(Box::new(generator));
@@ -473,9 +490,12 @@ impl TestCase {
 }
 
 /// TODO
-fn check(input: &str, output: &str, expected_diagnostics: &[Diagnostic],
-    constant_generator: Option<&dyn Fn(&InternPool) -> Vec<Value>>)
-{
+fn check(
+    input: &str,
+    output: &str,
+    expected_diagnostics: &[Diagnostic],
+    constant_generator: Option<&dyn Fn(&InternPool) -> Vec<Value>>,
+) {
     let pool = InternPool::new();
 
     let data = parse(&pool, input);
@@ -518,7 +538,8 @@ fn expand(pool: &InternPool, data: &[ScannedDatum]) -> Vec<Expression> {
     let (expansion_result, expansion_diagnostics) = collect_diagnostics(|handler| {
         let environment = basic_scheme_environment(pool);
 
-        return data.iter()
+        return data
+            .iter()
             .map(|datum| expand(datum, &environment, &handler))
             .collect();
     });
@@ -532,6 +553,6 @@ fn treat(expressions: &[Expression]) -> (MeaningResult, Vec<Diagnostic>) {
     use liblocus::utils::collect_diagnostics;
 
     collect_diagnostics(|handler| {
-        return meaning(handler, expressions);
+        return meaning(expressions, handler);
     })
 }
